# services/src/main/java/ru/relex/uber/services/model/UpdatableUser.java
package ru.relex.uber.services.model;

import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import org.immutables.value.Value;

@Value.Immutable
@JsonDeserialize(builder = ImmutableUpdatableUser.Builder.class)
public interface UpdatableUser extends BaseUser {

  String getPassword();

  UserStatus getStatus();
}